Added another condition in if statment

+++ all-Dokerfile/dokerscript.sh
@@ -26,21 +26,21 @@ echo "What is the port"
 read PORT
 
 #Write in the Dokerfile
-echo "FROM $IMAGE" >> Dockerfile
-echo "ENV MYSQL_ROOT_PASSWORD=$MYSQL_ROOT_PASSWORD" >> Dockerfile
-echo "ENV MYSQL_PASSWORD=$MYSQL_PASSWORD" >> Dockerfile
-echo "ENV MYSQL_USER=$MYSQL_USER" >> Dockerfile
-echo "ENV MYSQL_DATABASE=$MYSQL_DATABASE" >> Dockerfile
-echo "RUN mkdir -p /home/$WORKDIR" >> Dockerfile
-echo "COPY ./ /home/$WORKDIR" >> Dockerfile
-echo "WORKDIR /home/$WORKDIR" >> Dockerfile
-echo "EXPOSE $PORT " >> Dockerfile
-echo CMD ["apache2-foreground"] >> Dockerfile
+ "FROM $IMAGE" >> Dockerfile
+ "ENV MYSQL_ROOT_PASSWORD=$MYSQL_ROOT_PASSWORD" >> Dockerfile
+ "ENV MYSQL_PASSWORD=$MYSQL_PASSWORD" >> Dockerfile
+ "ENV MYSQL_USER=$MYSQL_USER" >> Dockerfile
+ "ENV MYSQL_DATABASE=$MYSQL_DATABASE" >> Dockerfile
+ "RUN mkdir -p /home/$WORKDIR" >> Dockerfile
+ "COPY ./ /home/$WORKDIR" >> Dockerfile
+ "WORKDIR /home/$WORKDIR" >> Dockerfile
+ "EXPOSE $PORT " >> Dockerfile
+ CMD ["apache2-foreground"] >> Dockerfile
 echo "Check this informations and answer y(Yes) if correct or no n(No)"
 cat Dockerfile
 read VALIDATOR
 val=y
-if [ "$VALIDATOR" ==  $val ]
+if [ -f Dockerfile && "$VALIDATOR" ==  $val ]
 then
     docker  build -f Dockerfile .
     rm -f Dockerfile
